refactor: extract wallet description component
