+++ packages/mask/src/plugins/Approval/SNSAdaptor/ApprovalDialog.tsx
@@ -4,14 +4,13 @@ import { TargetChainIdContext } from '@masknet/plugin-infra/web3-evm'
 import { useState } from 'react'
 import { NetworkTab } from '../../../components/shared/NetworkTab'
 import { useRemoteControlledDialog } from '@masknet/shared-base-ui'
-import { PluginWalletStatusBar } from '../../../utils/components/PluginWalletStatusBar'
 import { ChainId, chainResolver, NetworkType } from '@masknet/web3-shared-evm'
 import { NetworkPluginID } from '@masknet/web3-shared-base'
 import { useCurrentWeb3NetworkPluginID } from '@masknet/plugin-infra/web3'
 import { PluginId } from '@masknet/plugin-infra'
 import { useActivatedPlugin } from '@masknet/plugin-infra/dom'
 import { useI18N } from '../locales'
-import { useI18N as useBaseI18n } from '../../../utils'
+import { useI18N as useBaseI18n, PluginWalletStatusBar } from '../../../utils'
 import { WalletMessages } from '../../../plugins/Wallet/messages'
 import { InjectedDialog } from '@masknet/shared'
 import { useStyles } from './useStyles'
